# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PiperChain API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/wallets": {
      "post": {
        "tags": [
          "Wallets"
        ],
        "summary": "Create Wallet",
        "operationId": "create_wallet",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BasicResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/wallets/{address}/balance": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "Get Wallet Balance",
        "operationId": "get_wallet_balance",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Get Wallet Balance Wallets  Address  Balance Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/wallets/{address}/transactions": {
      "get": {
        "tags": [
          "Wallets"
        ],
        "summary": "Get Wallet Transactions",
        "operationId": "get_wallet_transactions",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionDetailResponse"
                  },
                  "title": "Response Get Wallet Transactions Wallets  Address  Transactions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BasicResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "BasicResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TransactionDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/TransactionInputDetail"
            },
            "type": "array",
            "title": "Inputs"
          },
          "outputs": {
            "items": {
              "$ref": "#/components/schemas/TransactionOutputDetail"
            },
            "type": "array",
            "title": "Outputs"
          },
          "fee": {
            "type": "integer",
            "title": "Fee"
          },
          "block_height": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Height"
          },
          "is_coinbase": {
            "type": "boolean",
            "title": "Is Coinbase"
          }
        },
        "type": "object",
        "required": [
          "id",
          "timestamp",
          "inputs",
          "outputs",
          "fee",
          "is_coinbase"
        ],
        "title": "TransactionDetailResponse"
      },
      "TransactionInputDetail": {
        "properties": {
          "ref_tx_id": {
            "type": "string",
            "title": "Ref Tx Id"
          },
          "ref_output_index": {
            "type": "integer",
            "title": "Ref Output Index"
          },
          "signature": {
            "type": "string",
            "title": "Signature"
          }
        },
        "type": "object",
        "required": [
          "ref_tx_id",
          "ref_output_index",
          "signature"
        ],
        "title": "TransactionInputDetail"
      },
...
      "TransactionOutputDetail": {
        "properties": {
          "recipient": {
            "type": "string",
            "title": "Recipient"
          },
          "amount": {
            "type": "integer",
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "recipient",
          "amount"
        ],
        "title": "TransactionOutputDetail"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WalletCreateRequest": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "address": {
            "type": "string",
            "title": "Address"
          },
          "public_key": {
            "type": "string",
            "title": "Public Key"
          },
          "encrypted_private_key": {
            "type": "string",
            "title": "Encrypted Private Key"
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "address",
          "public_key",
          "encrypted_private_key"
        ],
        "title": "WalletCreateRequest"
      }
    }
  }
}